Add sentiment analysis model and update gitignore
Added a LSTM sentiment analysis model implemented in Keras, that has been trained on a dataset. Added '.idea' to .gitignore. The model first preprocesses the text by tokenizing, removing stop words and lemmatizing, then use embedding and LSTM layers in the model. The model is trained, evaluated and predictions are shown in a confusion matrix.